% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{all_indicf}
\alias{all_indicf}
\title{Function that computes the spatial metrics}
\usage{
all_indicf(mat, N_SHUFFLE)
}
\arguments{
\item{mat}{}
}
\value{
all indic values
}
\description{
Function that computes the spatial metrics
}
\examples{
all_indicf(mat)
}
