% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{best_pl_or_tpl}
\alias{best_pl_or_tpl}
\title{Returns 1 if pl better, 0 if tpl better, NA if impossible fit}
\usage{
best_pl_or_tpl(mat)
}
\arguments{
\item{mat}{}
}
\value{
best fit
}
\description{
Returns 1 if pl better, 0 if tpl better, NA if impossible fit
}
\examples{
best_pl_or_tpl(mat)
}
