% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{fl_indic}
\alias{fl_indic}
\title{Takes the flowlength indicator. flowlength does not only depend on the matrix, but also on external variables (slope, cell_size). There is no way at the moment to pass those matrix-by-matrix, so we read them from attributes attached to the matrix.}
\usage{
fl_indic(m)
}
\arguments{
\item{m}{}
}
\value{
flowlength
}
\description{
Takes the flowlength indicator. flowlength does not only depend on the matrix, but also on external variables (slope, cell_size). There is no way at the moment to pass those matrix-by-matrix, so we read them from attributes attached to the matrix.
}
\examples{
fl_indic(m)
}
