% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compute_indicators.R
\name{function_compute_indicators}
\alias{function_compute_indicators}
\title{Function which computes all the indicators}
\usage{
function_compute_indicators(NPERM, N_SHUFFLE, path_output)
}
\arguments{
\item{NPERM}{}

\item{N_SHUFFLE}{}

\item{path_output}{}
}
\value{
indics
}
\description{
Function which computes all the indicators
}
\examples{
function_compute_indicators(NPERM,N_SHUFFLE,path_output)
}
