% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compute_indicators_data200.R
\name{function_compute_indicators_data200}
\alias{function_compute_indicators_data200}
\title{Function which computes all the spatial metrics in the 200 x 200 m images (subset of the Biocom data set)}
\usage{
function_compute_indicators_data200(NPERM, N_SHUFFLE, path_output)
}
\arguments{
\item{NPERM}{}

\item{path_output}{}
}
\value{
indics
}
\description{
Function which computes all the spatial metrics in the 200 x 200 m images (subset of the Biocom data set)
}
\examples{
function_compute_indicators_data200(NPERM,N_SHUFFLE,path_output)
}
