% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compute_indicators_data50.R
\name{function_compute_indicators_data50}
\alias{function_compute_indicators_data50}
\title{Function which computes all the spatial metrics in the 50 x 50 m images (three per site in 115 sites of the Biocom data set)}
\usage{
function_compute_indicators_data50(NPERM, N_SHUFFLE, path_output)
}
\arguments{
\item{path_output}{}
}
\value{
indics
}
\description{
Function which computes all the spatial metrics in the 50 x 50 m images (three per site in 115 sites of the Biocom data set)
}
\examples{
function_compute_indicators_data50(NPERM,N_SHUFFLE,path_output)
}
