% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compute_indicators_kefimodel_all.R
\name{function_compute_indicators_kefimodel_all}
\alias{function_compute_indicators_kefimodel_all}
\title{Function which computes all the spatial metrics on the outcomes of simulations from the model of Kéfi et al. 2007, Theoretical Population Biology}
\usage{
function_compute_indicators_kefimodel_all(
  NPERM,
  N_SHUFFLE,
  path_output,
  path_model_data
)
}
\arguments{
\item{path_model_data}{}
}
\value{
model_indics_df
}
\description{
Function which computes all the spatial metrics on the outcomes of simulations from the model of Kéfi et al. 2007, Theoretical Population Biology
}
\examples{
function_compute_indicators_kefimodel_all(NPERM,N_SHUFFLE,path_output,path_model_data)
}
