% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compute_indicators_scanlonmodel.R
\name{function_compute_indicators_scanlonmodel}
\alias{function_compute_indicators_scanlonmodel}
\title{Function which computes all the spatial metrics in the outcomes of simulations from the mode of Scanlon et al. 2007, Nature}
\usage{
function_compute_indicators_scanlonmodel(
  NPERM,
  N_SHUFFLE,
  path_output,
  path_model_data
)
}
\arguments{
\item{path_model_data}{}
}
\value{
model_indics_df
}
\description{
Function which computes all the spatial metrics in the outcomes of simulations from the mode of Scanlon et al. 2007, Nature
}
\examples{
function_compute_indicators_scanlonmodel(NPERM,N_SHUFFLE,path_output,path_model_data)
}
