% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_compute_model_indicators.R
\name{function_compute_model_indicators}
\alias{function_compute_model_indicators}
\title{Function which computes all the indicators}
\usage{
function_compute_model_indicators(NPERM, path_output, path_model_data)
}
\arguments{
\item{NPERM}{}

\item{path_output}{}

\item{path_model_data}{}
}
\value{
model_indics_df
}
\description{
Function which computes all the indicators
}
\examples{
function_compute_model_indicators(NPERM,path_output,path_model_data)
}
