% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_indicator_trends_data200.R
\name{function_indicator_trends_data200}
\alias{function_indicator_trends_data200}
\title{Function which estimates the trend (slope) of each of the spatial metrics along the aridity gradient in the 200 x 200 m images, in the whole data set or per group of sites}
\usage{
function_indicator_trends_data200(NPERM, N_SHUFFLE, path_output, BOOTN, ALPHA)
}
\arguments{
\item{NPERM}{}

\item{N_SHUFFLE}{}

\item{path_output}{}

\item{ALPHA}{}
}
\value{
trends
}
\description{
Function which estimates the trend (slope) of each of the spatial metrics along the aridity gradient in the 200 x 200 m images, in the whole data set or per group of sites
}
\examples{
function_indicator_trends_data200(NPERM,N_SHUFFLE,path_output,BOOTN,ALPHA)
}
