% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_indicator_trends_data50.R
\name{function_indicator_trends_data50}
\alias{function_indicator_trends_data50}
\title{Function which the trend (slope) of each of the spatial metrics along the aridity gradient in the 50 x 50 images, for the whole data set or per group of sites}
\usage{
function_indicator_trends_data50(NPERM, N_SHUFFLE, path_output, BOOTN, ALPHA)
}
\arguments{
\item{ALPHA}{}
}
\value{
trends
}
\description{
Function which the trend (slope) of each of the spatial metrics along the aridity gradient in the 50 x 50 images, for the whole data set or per group of sites
}
\examples{
function_indicator_trends_data50(NPERM,N_SHUFFLE,path_output,BOOTN,ALPHA)
}
