% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_indicator_trends_scanlonmodel.R
\name{function_indicator_trends_scanlonmodel}
\alias{function_indicator_trends_scanlonmodel}
\title{Function which the trend (slope) of each of the spatial metrics along the aridity gradient in the model}
\usage{
function_indicator_trends_scanlonmodel(
  NPERM,
  N_SHUFFLE,
  path_output,
  BOOTN,
  ALPHA
)
}
\arguments{
\item{NPERM}{}

\item{N_SHUFFLE}{}

\item{path_output}{}

\item{ALPHA}{}
}
\value{
trends
}
\description{
Function which the trend (slope) of each of the spatial metrics along the aridity gradient in the model
}
\examples{
function_indicator_trends_scanlonmodel(NPERM,N_SHUFFLE,path_output,BOOTN,ALPHA)
}
