% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_make_branches.R
\name{function_make_branches}
\alias{function_make_branches}
\title{Function which classifies sites into branches (or groups of dryland sites)}
\usage{
function_make_branches(NPERM, path_output)
}
\arguments{
\item{NPERM}{}

\item{path_output}{}
}
\value{
indics
}
\description{
Function which classifies sites into branches (or groups of dryland sites)
}
\examples{
function_make_branches(NPERM,path_output)
}
