% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_make_branches_data200.R
\name{function_make_branches_data200}
\alias{function_make_branches_data200}
\title{Function which includes the groups of dryland sites into the 200 x 200 data set}
\usage{
function_make_branches_data200(NPERM, path_output)
}
\arguments{
\item{NPERM}{}

\item{path_output}{}
}
\value{
indics
}
\description{
Function which includes the groups of dryland sites into the 200 x 200 data set
}
\examples{
function_make_branches_data200(NPERM,path_output)
}
