% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_make_branches_data50.R
\name{function_make_branches_data50}
\alias{function_make_branches_data50}
\title{Function which classifies sites into groups of dryland sites}
\usage{
function_make_branches_data50(NPERM, path_output)
}
\arguments{
\item{NPERM}{}

\item{path_output}{}
}
\value{
indics
}
\description{
Function which classifies sites into groups of dryland sites
}
\examples{
function_make_branches_data50(NPERM,path_output)
}
