% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_prepare_data.R
\name{function_prepare_data}
\alias{function_prepare_data}
\title{This file will read the biocom data from files and merge them into suitable tables for this analysis. Matrices are read and put in the same order as rows in the biocom database.}
\usage{
function_prepare_data(path_data, path_output)
}
\value{
ourdata.rda
}
\description{
This file will read the biocom data from files and merge them into suitable tables for this analysis. Matrices are read and put in the same order as rows in the biocom database.
}
\examples{
function_prepare_data(overwrite = FALSE)
}
