% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{get_ks_distance}
\alias{get_ks_distance}
\title{Computes the KS distance between the observed psd and the null psd, doing so multiple times to get a good estimate of the average ks distance to the null distribution.}
\usage{
get_ks_distance(mat, N_SHUFFLE)
}
\arguments{
\item{mat}{}
}
\value{
mean(all_ks)
}
\description{
Computes the KS distance between the observed psd and the null psd, doing so multiple times to get a good estimate of the average ks distance to the null distribution.
}
\examples{
get_ks_distance(mat)
}
