% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{get_lm}
\alias{get_lm}
\title{Get the slope of a set of samples contained in dat, identified by samples. The}
\usage{
get_lm(dat, samples)
}
\arguments{
\item{dat}{}

\item{samples}{}
}
\value{
a
}
\description{
Get the slope of a set of samples contained in dat, identified by samples. The
}
\examples{
get_lm(dat, samples)
}
