% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{get_lm_model}
\alias{get_lm_model}
\title{Function that estimates a slope on model data. There is no mixed effect because no need on model data, and we use aridity instead of Aridity because that is what is in the model data.We need to have the two arguments, dat and samples, because those are imposed by the boot() function given in base R.}
\usage{
get_lm_model(dat, samples)
}
\arguments{
\item{samples}{}
}
\value{
a\link{2}
}
\description{
Function that estimates a slope on model data. There is no mixed effect because no need on model data, and we use aridity instead of Aridity because that is what is in the model data.We need to have the two arguments, dat and samples, because those are imposed by the boot() function given in base R.
}
\examples{
get_lm_model(dat, samples)
}
