% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{get_slope_along_aridity_model}
\alias{get_slope_along_aridity_model}
\title{Function that given a dataset, will return BOOTN estimates of the slope along}
\usage{
get_slope_along_aridity_model(tbl, ...)
}
\arguments{
\item{tbl}{}

\item{...}{}
}
\description{
Function that given a dataset, will return BOOTN estimates of the slope along
}
\examples{
get_slope_along_aridity_model(tbl)
}
