% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{pvalstar}
\alias{pvalstar}
\title{Associates number of stars to a p-value}
\usage{
pvalstar(X, NAval = NA_character_)
}
\arguments{
\item{NAval}{}
}
\value{
a
}
\description{
Associates number of stars to a p-value
}
\examples{
pvalstar(X,NAval)
}
