% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{reset_workers}
\alias{reset_workers}
\title{Function that resets the workers before a parallel computation. It is
useful to do so just before we spawn something in parallel, because the
cluster of workers may sometimes hang.}
\usage{
reset_workers(workers = parallel::detectCores() - 1)
}
\arguments{
\item{workers}{The number of workers to use}
}
\description{
Function that resets the workers before a parallel computation. It is
useful to do so just before we spawn something in parallel, because the
cluster of workers may sometimes hang.
}
