% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{signf}
\alias{signf}
\title{Return the sign of something if significant. Used to set colors in graphs.}
\usage{
signf(x, alpha)
}
\arguments{
\item{alpha}{}
}
\value{
a
}
\description{
Return the sign of something if significant. Used to set colors in graphs.
}
\examples{
signf(x, alpha)
}
