% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialews_biocom-package.R
\docType{package}
\name{spatialews_biocom-package}
\alias{spatialews_biocom}
\alias{spatialews_biocom-package}
\title{spatialews_biocom: The Title of the Project}
\description{
A paragraph providing a full description of the project (on several lines...)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/skefi/spatialews_biocom}
  \item Report bugs at \url{https://github.com/skefi/spatialews_biocom/issues}
}

}
\author{
\strong{Maintainer}: Sonia Kefi \email{sonia.kefi@umontpellier.fr} (\href{https://orcid.org/0000-0002-9678-7770}{ORCID}) [copyright holder]

}
\keyword{internal}
