% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_helper.R
\name{stdz}
\alias{stdz}
\title{Standardize}
\usage{
stdz(
  X,
  mu = mean(X, na.rm = na.rm),
  sigma = sd(X, na.rm = na.rm),
  na.rm = FALSE
)
}
\arguments{
\item{X}{}

\item{mu}{}

\item{sigma}{}

\item{na.rm}{}
}
\value{
( X - mu ) / sigma
}
\description{
Standardize
}
\examples{
stdz()
}
